initialized frontend

+++ Backend/src/controllers/todo.controllers.js
@@ -1,20 +1,54 @@
+import Category from "../models/category.model.js";
 import Todo from "../models/todo.models.js";
 import asyncHandler from "../utils/asyncHandler.js";
 
-const addTodo = asyncHandler(async (req, res) => {
-    const { todoTitle, todoContent, category } = req.body;
-    if (!todoTitle || !todoContent || !category)
-        return res.sendError(400, "All fields required");
-    const existingTodo = await Todo.findOne({
+const createCategory = asyncHandler(async (req, res) => {
+    const { categoryName } = req.body;
+    if (!categoryName) return res.sendError(400, "Category name required");
+    const existingCategory = await Category.findOne({
         createdBy: req.user._id,
-        todoTitle,
-        category,
+        categoryName,
     });
-    if (existingTodo) return res.sendError(400, "same todo already exists");
+    if (existingCategory)
+        return res.sendError(400, "Same category already exists");
+
+    const category = await Category.create({
+        categoryName,
+        createdBy: req.user?._id,
+    });
+
+    return res.sendSuccess(200, category, "Category created successfully");
+});
+
+const deleteCategory = asyncHandler(async (req, res) => {
+    const { categoryId } = req.body;
+    const category = await Category.findOneAndDelete({ _id: categoryId });
+    if(!category) return res.sendError(400, "Invalid category id")
+    
+    await Todo.deleteMany({ _id: { $in: category.todos.map(todo => todo._id) } })
+    
+    return res.sendSuccess(200, "Category deleted successfully")
+});
+
+const addTodo = asyncHandler(async (req, res) => {
+    const { todoTitle, todoContent, categoryId } = req.body;
+    if (!todoTitle || !todoContent || !categoryId)
+        return res.sendError(400, "All fields required");
+
+    // const existingTodo = await Todo.findOne({
+    //     createdBy: req.user._id,
+    //     todoTitle
+    // });
+    const category = await Category.findOne({
+        _id: categoryId,
+        createdBy: req.user?._id,
+    });
+
+    if (!category) return res.sendError(400, "Invalid categoryId");
+
     const todo = await Todo.create({
         todoTitle,
         todoContent,
-        category,
         createdBy: req.user?._id,
     });
 
@@ -22,13 +56,26 @@ const addTodo = asyncHandler(async (req, res) => {
     if (!createdTodo) {
         throw new Error("Something went wrong while creating todo");
     }
+
+    category.todos.push(createdTodo._id);
+    await category.save();
+
     return res.sendSuccess(200, createdTodo, "Todo created successfully");
 });
 
 const deleteTodo = asyncHandler(async (req, res) => {
-    const { todoId } = req.body;
+    const { todoId, categoryId } = req.body;
+    const updatedCategory = await Category.findByIdAndUpdate(
+        categoryId,
+        {
+            $pull: { todos: { _id: todoId } },
+        },
+        { new: true }
+    );
+    if (!updatedCategory) return res.sendError(400, "Invalid categoryId");
     const deletedTodo = await Todo.findByIdAndDelete(todoId);
     if (!deletedTodo) return res.sendError(400, "Invalid todoId");
+
     return res.sendSuccess(200, {}, "Todo deleted successfully");
 });
 
@@ -47,4 +94,4 @@ const updateTodo = asyncHandler(async (req, res) => {
     return res.sendSuccess(200, updatedTodo, "Todo updated successfully");
 });
 
-export { addTodo, deleteTodo, updateTodo };
+export { addTodo, deleteTodo, updateTodo, createCategory, deleteCategory };
